+++ packages/ramps-controller/src/index.ts
@@ -18,7 +18,6 @@ export type {
   RampsServiceMessenger,
   Country,
   State,
-  Eligibility,
   CountryPhone,
   Provider,
   ProviderLink,
@@ -33,7 +32,6 @@ export {
 export type {
   RampsServiceGetGeolocationAction,
   RampsServiceGetCountriesAction,
-  RampsServiceGetEligibilityAction,
 } from './RampsService-method-action-types';
 export type {
   RequestCache,